job_portal_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
Pillow==10.1.0
whitenoise==6.6.0
requests==2.31.0
dj-database-url==2.1.0
